requirements.txt
python-dotenv~=1.0.1
langchain~=0.2.6
langchain-community
langchainhub
langchain-chroma
langchain-openai
bs4~=0.0.2
fastapi~=0.111.0
sqlmodel~=0.0.19
beautifulsoup4~=4.12.3
pypdf